Move the relay picker back into this crate:
I don't think it's used by others, it breaks often as we change nostr-types.

Also, I need to make changes and it is far harder as a separate crate as it isn't
really a clean interface (the interface may change marketdly and that would be a PITA
as a separate crate).