Check for and print if we have lock contention starting a relay, NOTES:
I did this because gossip was hanging after a while.  I've just run it again and it did
hang and there was never any lock contention here.  So this was not the problem.

It spins a feed waiting on ALL and ANY feed you check, and if you close down, it hangs
at: Overlord waiting for minions to all shutdown

^C doesn't recover.