Shows full path in case of a File Exists error
When the system tries to create the configuration directories,
it crashes if it fails (say because of regular file there or
a broken symlink). The out put is "File Exists" which makes
troubleshooting difficult.

With this change, gossip prints the full path of the
problematic file so the user khows where to look.

fixes #856