Undo Send for DMs too, but...:
We store delayed posts in a global array by Id.  But DMs are two events each.
It's clear when you undo a DM to get it's ID, but that is just your copy.
It is hard to associate the copy going to the counterparty.

So I just cleared the entire array of delayed posts.

The consequence is you can't have multiple posts in flight overlapping anymore
and undo them individually, because undo send now undoes ALL sends in progress.